integer for orders + auditing

+++ src/main/kotlin/com/fieldfreshmarket/api/controller/request/orders/sell/CreateSellProductRequest.kt
@@ -15,7 +15,7 @@ class CreateSellProductRequest(
     @get:NonNull
     val minPriceCents: Long,
     @get:NonNull
-    val volume: Double,
+    val volume: Int,
     @get:NonNull
     val productId: String,
     @get:NonNull
